ImageData classes now ready for common use.
