--- server/src/main/java/uk/ac/cam/echo/extrusionfinder/preprocessor/GrayscaleImageData.java
@@ -1,9 +1,0 @@
-// TEMPORARY!
-
-package uk.ac.cam.echo.extrusionfinder.preprocessor;
-
-public class GrayscaleImageData extends ImageData {
-    public GrayscaleImageData(byte[] data, int width, int height) {
-        super(data, width, height);
-    }
-}
