--- server/src/main/java/uk/ac/cam/echo/extrusionfinder/preprocessor/RGBImageData.java
@@ -1,9 +1,0 @@
-// TEMPORARY!
-
-package uk.ac.cam.echo.extrusionfinder.preprocessor;
-
-public class RGBImageData extends ImageData {
-    public RGBImageData(byte[] data, int width, int height) {
-        super(data, width, height);
-    }
-}
